#69: Tighten up Koinos Advantage section
